tools/postman_dependency_graph/requirements.txt
click>=8.0.0
python-dotenv>=0.19.0
networkx>=2.6.3
pyyaml>=6.0.1
